[fix](Nereids) type check could not work when root node is table or file sink
