+++ fe/fe-core/src/main/java/org/apache/doris/nereids/rules/RuleType.java
@@ -57,6 +57,7 @@ public enum RuleType {
     BINDING_SET_OPERATION_SLOT(RuleTypeClass.REWRITE),
     BINDING_GENERATE_FUNCTION(RuleTypeClass.REWRITE),
     BINDING_INSERT_TARGET_TABLE(RuleTypeClass.REWRITE),
+    BINDING_INSERT_FILE(RuleTypeClass.REWRITE),
 
     REPLACE_SORT_EXPRESSION_BY_CHILD_OUTPUT(RuleTypeClass.REWRITE),
