+++ fe/fe-core/src/main/java/org/apache/doris/nereids/trees/plans/logical/LogicalFileSink.java
@@ -19,15 +19,13 @@ package org.apache.doris.nereids.trees.plans.logical;
 
 import org.apache.doris.nereids.memo.GroupExpression;
 import org.apache.doris.nereids.properties.LogicalProperties;
-import org.apache.doris.nereids.trees.expressions.Expression;
-import org.apache.doris.nereids.trees.expressions.Slot;
+import org.apache.doris.nereids.trees.expressions.NamedExpression;
 import org.apache.doris.nereids.trees.plans.Plan;
 import org.apache.doris.nereids.trees.plans.PlanType;
 import org.apache.doris.nereids.trees.plans.algebra.Sink;
 import org.apache.doris.nereids.trees.plans.visitor.PlanVisitor;
 
 import com.google.common.base.Preconditions;
-import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
 
 import java.util.List;
@@ -44,33 +42,34 @@ public class LogicalFileSink<CHILD_TYPE extends Plan> extends LogicalSink<CHILD_
     private final String format;
     private final Map<String, String> properties;
 
-    public LogicalFileSink(String filePath, String format, Map<String, String> properties, CHILD_TYPE child) {
-        this(filePath, format, properties, Optional.empty(), Optional.empty(), child);
+    public LogicalFileSink(String filePath, String format,
+            Map<String, String> properties, List<NamedExpression> outputExprs, CHILD_TYPE child) {
+        this(filePath, format, properties, outputExprs, Optional.empty(), Optional.empty(), child);
     }
 
     public LogicalFileSink(String filePath, String format, Map<String, String> properties,
+            List<NamedExpression> outputExprs,
             Optional<GroupExpression> groupExpression, Optional<LogicalProperties> logicalProperties,
             CHILD_TYPE child) {
-        super(PlanType.LOGICAL_FILE_SINK, groupExpression, logicalProperties, child);
+        super(PlanType.LOGICAL_FILE_SINK, outputExprs, groupExpression, logicalProperties, child);
         this.filePath = Objects.requireNonNull(filePath);
         this.format = Objects.requireNonNull(format);
         this.properties = ImmutableMap.copyOf(Objects.requireNonNull(properties));
     }
 
+    public Plan withOutputExprs(List<NamedExpression> outputExprs) {
+        return new LogicalFileSink<>(filePath, format, properties, outputExprs, child());
+    }
+
     @Override
     public Plan withChildren(List<Plan> children) {
         Preconditions.checkArgument(children.size() == 1);
-        return new LogicalFileSink<>(filePath, format, properties, children.get(0));
+        return new LogicalFileSink<>(filePath, format, properties, outputExprs, children.get(0));
     }
 
     @Override
     public <R, C> R accept(PlanVisitor<R, C> visitor, C context) {
         return visitor.visitLogicalFileSink(this, context);
-    }
-
-    @Override
-    public List<? extends Expression> getExpressions() {
-        return ImmutableList.of();
     }
 
     @Override
@@ -81,33 +80,31 @@ public class LogicalFileSink<CHILD_TYPE extends Plan> extends LogicalSink<CHILD_
         if (o == null || getClass() != o.getClass()) {
             return false;
         }
+        if (!super.equals(o)) {
+            return false;
+        }
         LogicalFileSink<?> that = (LogicalFileSink<?>) o;
-        return Objects.equals(filePath, that.filePath)
-                && Objects.equals(format, that.format)
+        return Objects.equals(filePath, that.filePath) && Objects.equals(format, that.format)
                 && Objects.equals(properties, that.properties);
     }
 
     @Override
     public int hashCode() {
-        return Objects.hash(filePath, format, properties);
+        return Objects.hash(super.hashCode(), filePath, format, properties);
     }
 
     @Override
     public Plan withGroupExpression(Optional<GroupExpression> groupExpression) {
-        return new LogicalFileSink<>(filePath, format, properties, groupExpression,
-                Optional.of(getLogicalProperties()), child());
+        return new LogicalFileSink<>(filePath, format, properties, outputExprs,
+                groupExpression, Optional.of(getLogicalProperties()), child());
     }
 
     @Override
     public Plan withGroupExprLogicalPropChildren(Optional<GroupExpression> groupExpression,
             Optional<LogicalProperties> logicalProperties, List<Plan> children) {
         Preconditions.checkArgument(children.size() == 1);
-        return new LogicalFileSink<>(filePath, format, properties, groupExpression, logicalProperties, children.get(0));
-    }
-
-    @Override
-    public List<Slot> computeOutput() {
-        return child().getOutput();
+        return new LogicalFileSink<>(filePath, format, properties, outputExprs,
+                groupExpression, logicalProperties, children.get(0));
     }
 
     public String getFilePath() {
