+++ fe/fe-core/src/main/java/org/apache/doris/nereids/trees/plans/logical/LogicalResultSink.java
@@ -19,9 +19,7 @@ package org.apache.doris.nereids.trees.plans.logical;
 
 import org.apache.doris.nereids.memo.GroupExpression;
 import org.apache.doris.nereids.properties.LogicalProperties;
-import org.apache.doris.nereids.trees.expressions.Expression;
 import org.apache.doris.nereids.trees.expressions.NamedExpression;
-import org.apache.doris.nereids.trees.expressions.Slot;
 import org.apache.doris.nereids.trees.plans.Plan;
 import org.apache.doris.nereids.trees.plans.PlanType;
 import org.apache.doris.nereids.trees.plans.algebra.Sink;
@@ -29,10 +27,8 @@ import org.apache.doris.nereids.trees.plans.visitor.PlanVisitor;
 import org.apache.doris.nereids.util.Utils;
 
 import com.google.common.base.Preconditions;
-import com.google.common.collect.ImmutableList;
 
 import java.util.List;
-import java.util.Objects;
 import java.util.Optional;
 
 /**
@@ -40,22 +36,14 @@ import java.util.Optional;
  */
 public class LogicalResultSink<CHILD_TYPE extends Plan> extends LogicalSink<CHILD_TYPE> implements Sink {
 
-    private final List<NamedExpression> outputExprs;
-
     public LogicalResultSink(List<NamedExpression> outputExprs, CHILD_TYPE child) {
-        super(PlanType.LOGICAL_RESULT_SINK, child);
-        this.outputExprs = outputExprs;
+        super(PlanType.LOGICAL_RESULT_SINK, outputExprs, child);
     }
 
     public LogicalResultSink(List<NamedExpression> outputExprs,
             Optional<GroupExpression> groupExpression,
             Optional<LogicalProperties> logicalProperties, CHILD_TYPE child) {
-        super(PlanType.LOGICAL_RESULT_SINK, groupExpression, logicalProperties, child);
-        this.outputExprs = outputExprs;
-    }
-
-    public List<NamedExpression> getOutputExprs() {
-        return outputExprs;
+        super(PlanType.LOGICAL_RESULT_SINK, outputExprs, groupExpression, logicalProperties, child);
     }
 
     @Override
@@ -71,11 +59,6 @@ public class LogicalResultSink<CHILD_TYPE extends Plan> extends LogicalSink<CHIL
     }
 
     @Override
-    public List<? extends Expression> getExpressions() {
-        return outputExprs;
-    }
-
-    @Override
     public LogicalResultSink<Plan> withGroupExpression(Optional<GroupExpression> groupExpression) {
         return new LogicalResultSink<>(outputExprs, groupExpression, Optional.of(getLogicalProperties()), child());
     }
@@ -88,32 +71,8 @@ public class LogicalResultSink<CHILD_TYPE extends Plan> extends LogicalSink<CHIL
     }
 
     @Override
-    public List<Slot> computeOutput() {
-        return outputExprs.stream()
-                .map(NamedExpression::toSlot)
-                .collect(ImmutableList.toImmutableList());
-    }
-
-    @Override
     public String toString() {
         return Utils.toSqlString("LogicalResultSink[" + id.asInt() + "]",
                 "outputExprs", outputExprs);
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) {
-            return true;
-        }
-        if (o == null || getClass() != o.getClass()) {
-            return false;
-        }
-        LogicalResultSink<?> that = (LogicalResultSink<?>) o;
-        return Objects.equals(outputExprs, that.outputExprs);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(outputExprs);
     }
 }
